--- src/main/java/school/faang/user_service/adapter/UserRepositoryAdapter.java
@@ -1,17 +1,0 @@
-package school.faang.user_service.adapter;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import school.faang.user_service.entity.User;
-import school.faang.user_service.exception.ResourceNotFoundException;
-import school.faang.user_service.repository.UserRepository;
-
-@Component
-@RequiredArgsConstructor
-public class UserRepositoryAdapter {
-    private final UserRepository userRepository;
-
-    public User getById(Long id) {
-        return userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User not found by id: " + id));
-    }
-}
