+++ src/main/java/school/faang/user_service/repository/adapter/UserRepositoryAdapter.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 public class UserRepositoryAdapter {
     private final UserRepository userRepository;
 
-    public User findById(Long id) {
+    public User getById(Long id) {
         return userRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("User not found with id " + id));
     }
 }
